♻️  アイテム管理のロジックをコンポーザブルに移行し、ItemListコンポーネントでの状態管理を改善
